added localStorage function

+++ app.js
@@ -15,6 +15,38 @@ class Book {
 	}
 }
 
+// local storage class
+class LS {
+	static getBooks() {
+		//pull the books from the local storage
+		let books = JSON.parse(localStorage.getItem('books')) || [];
+		return books;
+	}
+	static displayBooks() {
+		// display the books from the local storage as soon the dom loaded
+		let books = LS.getBooks();
+		books.forEach((book) => {
+			ui.addBookToUi(book);
+		});
+	}
+	static addBooksToLS(book) {
+		//pull books from the LS, push the new book and reset the LS
+		let books = LS.getBooks();
+		books.push(book);
+		localStorage.setItem('books', JSON.stringify(books));
+	}
+	static deleteBookFromLS(isbnToRemove) {
+		let books = LS.getBooks();
+
+		books.forEach((book, index) => {
+			if (book.isbn === isbnToRemove) {
+				books.splice(index, 1);
+			}
+		});
+		localStorage.setItem('books', JSON.stringify(books));
+	}
+}
+
 // User interface class
 class UI {
 	addBookToUi(book) {
@@ -25,13 +57,9 @@ class UI {
         <td><a href="#"><i class="fas fa-trash delete__btn"></i></a></td>
 			`;
 		tbody.appendChild(tablerow);
-		this.showMessage('You book has been added successfully', 'green');
 	}
 	deleteBookFromUi(bookToDelete) {
-		if (confirm('Do you want to delete this book? ')) {
-			bookToDelete.remove();
-			this.showMessage('Book removed successfully', 'green');
-		}
+		bookToDelete.remove();
 	}
 	showMessage(msg, color) {
 		error_msg.textContent = msg;
@@ -49,8 +77,12 @@ class UI {
 	}
 }
 
-ui = new UI(); //create a new instance of UI
+document.addEventListener('DOMContentLoaded', () => {
+	LS.displayBooks();
+	//pull books from LS as soon as the document loaded and display
+});
 
+const ui = new UI(); //create a new instance of UI
 form.addEventListener('submit', (e) => {
 	let title_value = title.value.trim(),
 		author_value = author.value.trim(),
@@ -59,8 +91,10 @@ form.addEventListener('submit', (e) => {
 	if (title_value === '' || author_value === '' || isbn_value === '') {
 		ui.showMessage('Please fill the complete information', 'red');
 	} else {
-		book = new Book(title_value, author_value, isbn_value);
+		const book = new Book(title_value, author_value, isbn_value);
 		ui.addBookToUi(book);
+		LS.addBooksToLS(book); //because it is static method, no need to instantiate
+		ui.showMessage('You book has been added successfully', 'green');
 		ui.clearText();
 	}
 	e.preventDefault();
@@ -73,6 +107,12 @@ tbody.addEventListener('click', (e) => {
 	if (e.target.classList.contains('delete__btn')) {
 		//go from i tag to a tag to td then to tr and delete the tr;
 		let element = e.target.parentElement.parentElement.parentElement;
-		ui.deleteBookFromUi(element);
+		if (confirm('Do you want to delete this book? ')) {
+			ui.deleteBookFromUi(element);
+			// to remove book from local storage, we target the td isbn number since they will be unique which is i to a to td and td previouselementsibling
+			let content = e.target.parentElement.parentElement.previousElementSibling.textContent;
+			LS.deleteBookFromLS(content);
+			ui.showMessage('Book removed successfully', 'green');
+		}
 	}
 });
